fix: 홈페이지 ssr로 변경
